refactor: accumulate the neutral scope in integers
Its density cloud carried floats and its average carried doubles, and neither can
be merged bit-exactly by a pass split across chunks - float addition is not
associative, so the result would depend on how many chunks the machine chose, and
the exact scope goldens run on three runners with different core counts.

The cloud holds quarter units instead, which is what the splat's own weights
already are: a whole sample on the centre cell, a half on each edge, a quarter on
each corner. Nothing about the drawing changes, because a cell and the peak it is
scaled against move together.

The average accumulates a twelve-bit fixed point. That quantizes each sample's
chroma by one part in four thousand - four orders of magnitude below the movement
the sampling budget already permits - and makes the sums whole numbers that add in
any order.

No behaviour changes here; this is what the parallel pass needs to exist.

+++ src/core/scopes/neutral.h
@@ -125,7 +125,11 @@ private:
     NeutralSettings m_settings;
     int m_imageSize = DefaultNeutralSize;
     float m_axisRange = 40.0f;
-    std::vector<float> m_cloud;  // m_imageSize x m_imageSize near-neutral density
+    // m_imageSize x m_imageSize near-neutral density, in quarter units so the
+    // splat's 1, 1/2 and 1/4 weights are whole numbers. Integers because a
+    // chunked accumulate has to merge bit-exactly however many chunks it used,
+    // and float addition is not associative.
+    std::vector<uint32_t> m_cloud;
     NormalizedPoint m_average{0.5f, 0.5f};
     uint64_t m_neutralCount = 0;
     bool m_hasData = false;
